Enable ruby installation to be configured
Add --with-ruby-install-dir and --with-ruby-install-dir-variable to
configure.  The former allows any installation path to be used, the
latter sets the configuration variable to use, defaulting to
'archdir' (the current config).

Replaced rubyarch_dir with rubyinstall_dir in Makefile.am for ruby
bindings since it may or may not be the Ruby 'archdir'.

Added documentation for the new options

Fixes Issue#0000412
http://bugs.librdf.org/mantis/view.php?id=412